userapp/requirements.txt
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4

sqlparse==0.4.1

